backend/requirements.txt
asgiref==3.7.2
attrs==23.2.0
Django==5.0.1
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
psycopg2==2.9.9
PyJWT==2.8.0
python-decouple==3.8
pytz==2023.4
PyYAML==6.0.1
referencing==0.33.0
rpds-py==0.17.1
sqlparse==0.4.4
tzdata==2023.4
uritemplate==4.1.1
